Fix #1 - #4 && created build file
Fix #1: moveBy center.y was not being added by the yOffset. This made the simpleMove() fail when y added was different from x added.

Fix #2: scale() - the radius was added to the factor being scaled, not multiplied. This made enlarge() fail

Fix #3: x & y values were flipped in Circle2 constructor. This made simpleMove() fail when trying to move the circle.

Fix#4: The intersect() of Circle2  now takes into account the sum of the radii in comparison to the distance of the two centers of the circle.